correct type for swr memory and refactor
